requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
blis==1.3.3
cachetools==7.0.5
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
gitdb==4.0.12
gitpython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
narwhals==2.18.0
networkx==3.6.1
numpy==2.4.3
opencv-contrib-python==4.13.0.92
packaging==26.0
paho-mqtt==2.1.0
pandas==2.3.3
pillow==12.1.1
preshed==3.0.12
proto-plus==1.27.1
protobuf==5.29.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydeck==0.9.1
pygments==2.19.2
pyparsing==3.3.2
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart-open==7.5.1
smmap==5.0.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2

tenacity==9.1.4
thinc==8.3.10
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.3
wrapt==2.1.2
python-multipart==0.0.9
